вынесен прескоринг к ласс PrescoringService, ConveyorApplication в корне проекта
