+++ src/main/java/com/CreditPipeline/conveyor/enums/Gender.java
@@ -1,4 +1,4 @@
-package enums;
+package com.CreditPipeline.conveyor.enums;
 
 public enum Gender {
     Male,
