+++ src/main/java/com/CreditPipeline/conveyor/enums/MaritalStatus.java
@@ -1,4 +1,4 @@
-package enums;
+package com.CreditPipeline.conveyor.enums;
 
 public enum MaritalStatus {
     Married,
