+++ src/main/java/com/CreditPipeline/conveyor/enums/Position.java
@@ -1,4 +1,4 @@
-package enums;
+package com.CreditPipeline.conveyor.enums;
 
 public enum Position {
     Manager,
