+++ src/main/java/com/CreditPipeline/conveyor/swagger/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package swagger;
+package com.CreditPipeline.conveyor.swagger;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
