+++ src/test/test/controller/OffersTest.java
@@ -1,8 +1,8 @@
 package controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
-import DTO.LoanOfferDTO;
+import com.CreditPipeline.conveyor.DTO.LoanOfferDTO;
+import com.CreditPipeline.conveyor.controller.Offers;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
